refactor: move implementations to the resources folder
