+++ pulumi/resources/LambdaApiGatewayV2.ts
@@ -1,6 +1,6 @@
 import * as aws from "@pulumi/aws";
 import * as pulumi from "@pulumi/pulumi";
-import { ifObjectEmpty } from "../src/utils/transformers";
+import { ifObjectEmpty } from "../../src/utils/transformers";
 
 /* ---------------Types------------------- */
